moved over all archive related functions and routes
