+++ server/server/routing/db_routes.js
@@ -40,11 +40,11 @@ const DB_CONFIG = require("../database/db_config");
 const CONFIG = require("../config/config");
 const { nanoid } = require("nanoid");
 // Split route modules (kept mounted here so existing API paths remain available)
-const sponsorsRoutes = require("./db_routes_parts/sponsors_routes");
-const actionsRoutes = require("./db_routes_parts/actions_routes");
-const projectsRoutes = require("./db_routes_parts/projects_routes");
-const archivesRoutes = require("./db_routes_parts/archives_routes");
-const errorLogsRoutes = require("./db_routes_parts/error_logs_routes");
+const sponsorsRoutes = require("./db_ind_routes/sponsors_routes");
+const actionsRoutes = require("./db_ind_routes/actions_routes");
+const projectsRoutes = require("./db_ind_routes/projects_routes");
+const archivesRoutes = require("./db_ind_routes/archives_routes");
+const errorLogsRoutes = require("./db_ind_routes/error_logs_routes");
 
 // Mount split routers early so they take precedence over the large monolithic definitions
 db_router.use("/", sponsorsRoutes(db));
@@ -572,1126 +572,6 @@ module.exports = (db) => {
         .catch((err) => {
           console.error(err);
           let error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getActiveProjects",
-    [UserAuth.isSignedIn],
-    (req, res, next) => {
-      let getProjectsQuery = `
-            SELECT *
-            FROM projects
-            LEFT JOIN semester_group
-            ON projects.semester = semester_group.semester_id
-            WHERE projects.semester IS NOT NULL
-        `;
-      db.query(getProjectsQuery)
-        .then((values) => {
-          res.send(values);
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getActiveCoaches",
-    [UserAuth.isCoachOrAdmin],
-    (req, res, next) => {
-      const sql = `SELECT * FROM users WHERE type = '${ROLES.COACH}' AND active = ''`;
-
-      db.query(sql)
-        .then((coaches) => {
-          res.send(coaches);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getProjectCoaches",
-    [UserAuth.isCoachOrAdmin],
-    (req, res, next) => {
-      const getProjectCoaches = `SELECT users.* FROM users
-            LEFT JOIN project_coaches ON project_coaches.coach_id = users.system_id
-            WHERE project_coaches.project_id = ?`;
-
-      db.query(getProjectCoaches, [req.query.project_id])
-        .then((coaches) => {
-          res.send(coaches);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getProjectStudents",
-    [UserAuth.isCoachOrAdmin],
-    (req, res, next) => {
-      const getProjectStudents = "SELECT * FROM users WHERE users.project = ?";
-
-      db.query(getProjectStudents, [req.query.project_id])
-        .then((students) => {
-          res.send(students);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getProjectStudentNames",
-    [UserAuth.isSignedIn],
-    (req, res, next) => {
-      const getProjectStudents =
-        "SELECT fname,lname FROM users WHERE users.project = ? and users.system_id!=?";
-
-      db.query(getProjectStudents, [req.query.project_id, req.user.system_id])
-        .then((students) => {
-          res.send(students);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/selectAllCoachInfo",
-    [UserAuth.isCoachOrAdmin],
-    (req, res, next) => {
-      const getCoachInfoQuery = `
-            SELECT users.system_id,
-            users.fname,
-            users.lname,
-            users.email,
-            users.semester_group,
-            (
-                SELECT "[" || group_concat(
-                    "{" ||
-                        """title"""         || ":" || """" || COALESCE(projects.display_name, projects.title) || """" || "," ||
-                        """semester_id"""   || ":" || """" || projects.semester                               || """" || "," ||
-                        """project_id"""    || ":" || """" || projects.project_id                             || """" || "," ||
-                        """organization"""  || ":" || """" || projects.organization                           || """" || "," ||
-                        """status"""        || ":" || """" || projects.status                                 || """" ||
-                    "}"
-                ) || "]"
-                FROM project_coaches
-                LEFT JOIN projects ON projects.project_id = project_coaches.project_id
-                WHERE project_coaches.coach_id = users.system_id
-            ) projects
-            FROM users
-            WHERE users.type = "${ACTION_TARGETS.COACH}"
-        `;
-
-      db.query(getCoachInfoQuery)
-        .then((coaches) => {
-          res.send(coaches);
-        })
-        .catch((err) => {
-          console.error(error);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  // used in the /projects page and home page if featured
-  db_router.get("/getActiveArchiveProjects", (req, res, next) => {
-    const { resultLimit, page, featured } = req.query;
-    let skipNum = page * resultLimit;
-    let projectsQuery;
-    let rowCountQuery;
-    if (featured === "true") {
-      // home page - randomized order of projects
-      projectsQuery = `SELECT * FROM ${DB_CONFIG.tableNames.archive} WHERE oid NOT IN
-            ( SELECT oid FROM ${DB_CONFIG.tableNames.archive} ORDER BY random() LIMIT ? )
-            AND inactive = '' AND featured = 1 ORDER BY random() LIMIT ?`;
-      rowCountQuery = `SELECT COUNT(*) FROM ${DB_CONFIG.tableNames.archive} WHERE inactive = ''`;
-    } else {
-      // projects page - all archived projects data regardless if they are archived or not
-      projectsQuery = `SELECT * FROM ${DB_CONFIG.tableNames.archive} WHERE oid NOT IN
-            ( SELECT oid FROM ${DB_CONFIG.tableNames.archive} ORDER BY archive_id LIMIT ? )
-            AND inactive = '' ORDER BY archive_id LIMIT ?`;
-      rowCountQuery = `SELECT COUNT(*) FROM ${DB_CONFIG.tableNames.archive} WHERE inactive = ''`;
-    }
-    const projectsPromise = db.query(projectsQuery, [skipNum, resultLimit]);
-    const rowCountPromise = db.query(rowCountQuery);
-    Promise.all([rowCountPromise, projectsPromise])
-      .then(([[rowCount], projects]) => {
-        res.send({
-          totalProjects: rowCount[Object.keys(rowCount)[0]],
-          projects: projects,
-        });
-      })
-      .catch((err) => {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  // endpoint for getting ALL archive data to view within admin view/editor
-  db_router.get("/getArchiveProjects", (req, res, next) => {
-    const { resultLimit, offset } = req.query;
-    let skipNum = offset * resultLimit;
-    let projectsQuery = `SELECT * FROM ${DB_CONFIG.tableNames.archive} WHERE
-            oid NOT IN (SELECT oid FROM ${DB_CONFIG.tableNames.archive} ORDER BY archive_id LIMIT ?)
-            ORDER BY archive_id LIMIT ?`;
-    let rowCountQuery = `SELECT COUNT(*) FROM ${DB_CONFIG.tableNames.archive}`;
-
-    const projectsPromise = db.query(projectsQuery, [skipNum, resultLimit]);
-    const rowCountPromise = db.query(rowCountQuery);
-
-    Promise.all([rowCountPromise, projectsPromise])
-      .then(([[rowCount], projects]) => {
-        res.send({
-          totalProjects: rowCount[Object.keys(rowCount)[0]],
-          projects: projects,
-        });
-      })
-      .catch((err) => {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  /**
-   * Responds with projects from database
-   *
-   * TODO: Add pagination
-   */
-  db_router.get(
-    "/getProjects",
-    [UserAuth.isCoachOrAdmin],
-    async (req, res, next) => {
-      const query = "SELECT * from projects";
-      db.query(query)
-        .then((projects) => res.send(projects))
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getCandidateProjects",
-    [UserAuth.isSignedIn],
-    async (req, res, next) => {
-      const query =
-        "SELECT * from projects WHERE projects.status = 'candidate';";
-      db.query(query)
-        .then((projects) => res.send(projects))
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.get(
-    "/getMyProjects",
-    [UserAuth.isSignedIn],
-    async (req, res, next) => {
-      let query;
-      let params;
-      switch (req.user.type) {
-        case ROLES.COACH:
-          query = `SELECT projects.*
-                FROM projects
-                INNER JOIN project_coaches
-                ON (projects.project_id = project_coaches.project_id AND project_coaches.coach_id = ?);`;
-          params = [req.user.system_id];
-          break;
-        case ROLES.STUDENT:
-          query = `SELECT users.system_id, users.semester_group, projects.*
-                FROM users
-                INNER JOIN projects
-                ON users.system_id = ? AND projects.project_id = users.project;`;
-          params = [req.user.system_id];
-          break;
-        case ROLES.ADMIN:
-          query =
-            "SELECT * FROM projects WHERE projects.status NOT IN ('completed', 'rejected', 'archive');";
-          params = [];
-          break;
-        default:
-          const error = new Error(
-            "Invalid user type...something must be very very broken...",
-          );
-          error.statusCode = 500;
-          return next(error);
-      }
-
-      db.query(query, params)
-        .then((proposals) => res.send(proposals))
-        .catch((err) => res.status(500).send(err));
-    },
-  );
-
-  db_router.get(
-    "/getSemesterProjects",
-    [UserAuth.isSignedIn],
-    async (req, res, next) => {
-      let query;
-      let params;
-      switch (req.user.type) {
-        case ROLES.COACH:
-          query = `
-                SELECT projects.*
-                FROM projects
-                WHERE projects.semester IN
-                    (SELECT projects.semester
-                    FROM projects
-                    INNER JOIN project_coaches
-                    ON (projects.project_id = project_coaches.project_id AND project_coaches.coach_id = ?))
-                ;`;
-          params = [req.user.system_id];
-          break;
-        case ROLES.STUDENT:
-          query = `SELECT users.system_id, projects.*
-                FROM users
-                INNER JOIN projects
-                ON users.system_id = ? AND projects.semester = users.semester_group;`;
-          params = [req.user.system_id];
-          break;
-        case ROLES.ADMIN:
-          query =
-            "SELECT * FROM projects WHERE projects.status NOT IN ('in progress', 'completed', 'rejected', 'archive');";
-          params = [];
-          break;
-        default:
-          const error = new Error(
-            "Invalid user type...something must be very very broken...",
-          );
-          error.statusCode = 500;
-          return next(error);
-      }
-
-      db.query(query, params)
-        .then((projects) => res.send(projects))
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/editArchive",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    async (req, res, next) => {
-      let body = req.body;
-      const updateArchiveQuery = `UPDATE ${DB_CONFIG.tableNames.archive}
-                                    SET featured=?, outstanding=?, creative=?, priority=?,
-                                        title=?, project_id=?, team_name=?,
-                                        members=?, sponsor=?, coach=?,
-                                        poster_thumb=?, poster_full=?, archive_image=?, synopsis=?,
-                                        video=?, name=?, dept=?,
-                                        start_date=?, end_date=?, keywords=?, url_slug=?, inactive=?, locked=?
-                                    WHERE archive_id = ?`;
-      const inactive =
-        body.inactive === "true"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const locked =
-        body.locked === "true"
-          ? req.user.fname +
-            " " +
-            req.user.lname +
-            " locked at " +
-            moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const checkBox = (data) => {
-        if (data === "true" || data === "1") {
-          return 1;
-        }
-        return 0;
-      };
-
-      const strToInt = (data) => {
-        if (typeof data === "string") {
-          return parseInt(data);
-        }
-        return 0;
-      };
-
-      let updateArchiveParams = [
-        checkBox(body.featured),
-        checkBox(body.outstanding),
-        checkBox(body.creative),
-        strToInt(body.priority),
-        body.title,
-        body.project_id,
-        body.team_name,
-        body.members,
-        body.sponsor,
-        body.coach,
-        body.poster_thumb,
-        body.poster_full,
-        body.archive_image,
-        body.synopsis,
-        body.video,
-        body.name,
-        body.dept,
-        body.start_date,
-        body.end_date,
-        body.keywords,
-        body.url_slug,
-        inactive,
-        locked,
-        body.archive_id,
-      ];
-
-      db.query(updateArchiveQuery, updateArchiveParams)
-        .then(() => {
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/createArchive",
-    [UserAuth.isAdmin, UserAuth.canWrite],
-    body("featured")
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-      .withMessage("Cannot be empty"),
-    async (req, res, next) => {
-      let body = req.body;
-      const inactive =
-        body.inactive === "true"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-      const locked =
-        body.locked === "true"
-          ? req.user.fname +
-            " " +
-            req.user.lname +
-            " locked at " +
-            moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const updateArchiveQuery = `INSERT INTO ${DB_CONFIG.tableNames.archive}(featured, outstanding, creative,
-                                    priority, title, project_id, team_name, members, sponsor, coach, poster_thumb,
-                                    poster_full, archive_image, synopsis, video, name, dept, start_date, end_date,
-                                    keywords, url_slug, inactive, locked)
-                                    VALUES(?, ?, ?, ?, ?, ?, ?, ?,
-                                           ?, ?, ?, ?, ?, ?, ?, ?,
-                                           ?, ?, ?, ?, ?, ?, ?);`;
-
-      const checkBox = (data) => {
-        if (data === "true" || data === "1") {
-          return 1;
-        }
-        return 0;
-      };
-
-      const strToInt = (data) => {
-        if (typeof data === "string") {
-          return parseInt(data);
-        }
-        return 0;
-      };
-
-      const updateArchiveParams = [
-        checkBox(body.featured),
-        checkBox(body.outstanding),
-        checkBox(body.creative),
-        strToInt(body.priority),
-        body.title,
-        body.project_id,
-        body.team_name,
-        body.members,
-        body.sponsor,
-        body.coach,
-        body.poster_thumb,
-        body.poster_full,
-        body.archive_image,
-        body.synopsis,
-        body.video,
-        body.name,
-        body.dept,
-        body.start_date,
-        body.end_date,
-        body.keywords,
-        body.url_slug,
-        inactive,
-        locked,
-      ];
-
-      db.query(updateArchiveQuery, updateArchiveParams)
-        .then((response) => {
-          return res.status(200).send(response);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/editArchiveStudent",
-    [UserAuth.isSignedIn, UserAuth.canWrite],
-    async (req, res, next) => {
-      let body = req.body;
-      let files = req.files;
-      const updateArchiveQuery = `UPDATE ${DB_CONFIG.tableNames.archive}
-                                    SET featured=?, outstanding=?, creative=?, priority=?,
-                                        title=?, project_id=?, team_name=?,
-                                        members=?, sponsor=?, coach=?,
-                                        poster_thumb=?, poster_full=?, archive_image=?, synopsis=?,
-                                        video=?, name=?, dept=?,
-                                        start_date=?, end_date=?, keywords=?, url_slug=?, inactive=?, locked=?
-                                    WHERE archive_id = ?`;
-      const inactive =
-        body.inactive === "true"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const locked =
-        body.locked === "true"
-          ? req.user.fname +
-            " " +
-            req.user.lname +
-            " locked at " +
-            moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      let files_uploaded = [];
-
-      let poster_full = ``;
-      let poster_thumb = ``;
-      let archive_image = ``;
-      let video = ``;
-      if (!(files === undefined || files === null)) {
-        if (files.poster_full === undefined) {
-          poster_full = body.poster_full;
-          poster_thumb = body.poster_thumb;
-        } else {
-          if (
-            files.poster_full.mimetype == "image/png" &&
-            files.poster_full.size <= 30000000
-          ) {
-            files.poster_full.name = body.url_slug + "-poster";
-            poster_full = `${files.poster_full.name}`;
-            poster_thumb = poster_full;
-            let poster_URL = path.join(
-              __dirname,
-              `../../resource/archivePosters`,
-            );
-            files_uploaded.push([files.poster_full, poster_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        if (files.archive_image === undefined) {
-          archive_image = body.archive_image;
-        } else {
-          if (
-            files.archive_image.mimetype == "image/png" &&
-            files.archive_image.size <= 30000000
-          ) {
-            files.archive_image.name = body.url_slug + "-image";
-            archive_image = `${files.archive_image.name}`;
-            let image_URL = path.join(
-              __dirname,
-              `../../resource/archiveImages`,
-            );
-            files_uploaded.push([files.archive_image, image_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        if (files.video === undefined) {
-          video = body.video;
-        } else {
-          if (
-            files.video.mimetype == "video/mp4" &&
-            files.video.size <= 300000000
-          ) {
-            files.video.name = body.url_slug + "-video";
-            video = `${files.video.name}`;
-            let video_URL = path.join(
-              __dirname,
-              `../../resource/archiveVideos`,
-            );
-            files_uploaded.push([files.video, video_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        for (let i = 0; i < files_uploaded.length; i++) {
-          fs.mkdirSync(files_uploaded[i][1], { recursive: true });
-          files_uploaded[i][0].mv(
-            `${files_uploaded[i][1]}/${files_uploaded[i][0].name}`,
-            function (err) {
-              if (err) {
-                const error = new Error(err);
-                error.statusCode = 500;
-                return next(error);
-              }
-            },
-          );
-        }
-      } else {
-        poster_full = body.poster_full;
-        poster_thumb = body.poster_thumb;
-        archive_image = body.archive_image;
-        video = body.video;
-      }
-
-      const checkBox = (data) => {
-        if (data === "true" || data === "1") {
-          return 1;
-        }
-        return 0;
-      };
-
-      const strToInt = (data) => {
-        if (typeof data === "string") {
-          return parseInt(data);
-        }
-        return 0;
-      };
-
-      let updateArchiveParams = [
-        checkBox(body.featured),
-        checkBox(body.outstanding),
-        checkBox(body.creative),
-        strToInt(body.priority),
-        body.title,
-        body.project_id,
-        body.team_name,
-        body.members,
-        body.sponsor,
-        body.coach,
-        poster_thumb,
-        poster_full,
-        archive_image,
-        body.synopsis,
-        video,
-        body.name,
-        body.dept,
-        body.start_date,
-        body.end_date,
-        body.keywords,
-        body.url_slug,
-        inactive,
-        locked,
-        body.archive_id,
-      ];
-
-      db.query(updateArchiveQuery, updateArchiveParams)
-        .then(() => {
-          return res.status(200).send();
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  db_router.post(
-    "/createArchiveStudent",
-    [UserAuth.isSignedIn, UserAuth.canWrite],
-    body("featured")
-      .not()
-      .isEmpty()
-      .trim()
-      .escape()
-      .withMessage("Cannot be empty"),
-    async (req, res, next) => {
-      let body = req.body;
-      const inactive =
-        body.inactive === "true"
-          ? moment().format(CONSTANTS.datetime_format)
-          : "";
-      const locked =
-        body.locked === "true"
-          ? req.user.fname +
-            " " +
-            req.user.lname +
-            " locked at " +
-            moment().format(CONSTANTS.datetime_format)
-          : "";
-
-      const name = body.url_slug; //this value needs to be unique, but isn't used, so this is a relatively safe method.
-
-      let files = req.files;
-      let files_uploaded = [];
-
-      let poster_full = ``;
-      let poster_thumb = ``;
-      let archive_image = ``;
-      let video = ``;
-      if (!(files === undefined || files === null)) {
-        if (files.poster_full === undefined) {
-          poster_full = body.poster_full;
-          poster_thumb = body.poster_thumb;
-        } else {
-          if (
-            files.poster_full.mimetype == "image/png" &&
-            files.poster_full.size <= 30000000
-          ) {
-            files.poster_full.name = body.url_slug + "-poster";
-            poster_full = `${files.poster_full.name}`;
-            poster_thumb = poster_full;
-            let poster_URL = path.join(
-              __dirname,
-              `../../resource/archivePosters`,
-            );
-            files_uploaded.push([files.poster_full, poster_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        if (files.archive_image === undefined) {
-          archive_image = body.archive_image;
-        } else {
-          if (
-            files.archive_image.mimetype == "image/png" &&
-            files.archive_image.size <= 30000000
-          ) {
-            files.archive_image.name = body.url_slug + "-image";
-            archive_image = `${files.archive_image.name}`;
-            let image_URL = path.join(
-              __dirname,
-              `../../resource/archiveImages`,
-            );
-            files_uploaded.push([files.archive_image, image_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        if (files.video === undefined) {
-          video = body.video;
-        } else {
-          if (
-            files.video.mimetype == "video/mp4" &&
-            files.video.size <= 300000000
-          ) {
-            files.video.name = body.url_slug + "-video";
-            video = `${files.video.name}`;
-            let video_URL = path.join(
-              __dirname,
-              `../../resource/archiveVideos`,
-            );
-            files_uploaded.push([files.video, video_URL]);
-          } else {
-            const error = new Error(err);
-            error.statusCode = 500;
-            return next(error);
-          }
-        }
-
-        for (let i = 0; i < files_uploaded.length; i++) {
-          fs.mkdirSync(files_uploaded[i][1], { recursive: true });
-          if (
-            fs.existsSync(
-              `${files_uploaded[i][1]}/${files_uploaded[i][0].name}`,
-            )
-          ) {
-            fs.unlink(`${files_uploaded[i][1]}/${files_uploaded[i][0].name}`);
-          }
-          files_uploaded[i][0].mv(
-            `${files_uploaded[i][1]}/${files_uploaded[i][0].name}`,
-            function (err) {
-              if (err) {
-                const error = new Error(err);
-                error.statusCode = 500;
-                return next(error);
-              }
-            },
-          );
-        }
-      } else {
-        poster_full = body.poster_full;
-        poster_thumb = body.poster_thumb;
-        archive_image = body.archive_image;
-        video = body.video;
-      }
-
-      const updateArchiveQuery = `INSERT INTO ${DB_CONFIG.tableNames.archive}(featured, outstanding, creative,
-                                    priority, title, project_id, team_name, members, sponsor, coach, poster_thumb,
-                                    poster_full, archive_image, synopsis, video, name, dept, start_date, end_date,
-                                    keywords, url_slug, inactive, locked)
-                                    VALUES(?, ?, ?, ?, ?, ?, ?, ?,
-                                           ?, ?, ?, ?, ?, ?, ?, ?,
-                                           ?, ?, ?, ?, ?, ?, ?);`;
-
-      const checkBox = (data) => {
-        if (data === "true" || data === "1") {
-          return 1;
-        }
-        return 0;
-      };
-
-      const strToInt = (data) => {
-        if (typeof data === "string") {
-          return parseInt(data);
-        }
-        return 0;
-      };
-
-      const updateArchiveParams = [
-        checkBox(body.featured),
-        checkBox(body.outstanding),
-        checkBox(body.creative),
-        strToInt(body.priority),
-        body.title,
-        body.project_id,
-        body.team_name,
-        body.members,
-        body.sponsor,
-        body.coach,
-        poster_thumb,
-        poster_full,
-        archive_image,
-        body.synopsis,
-        video,
-        name,
-        body.dept,
-        body.start_date,
-        body.end_date,
-        body.keywords,
-        body.url_slug,
-        inactive,
-        locked,
-      ];
-
-      db.query(updateArchiveQuery, updateArchiveParams)
-        .then((response) => {
-          return res.status(200).send(response);
-        })
-        .catch((err) => {
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  //Gets the start and end dates of a project based on the semester that it is associated with.
-  db_router.get("/getProjectDates", UserAuth.isSignedIn, (req, res, next) => {
-    const getDatesQuery = `SELECT start_date, end_date FROM semester_group WHERE semester_id = ?`;
-    const getDatesParams = [req.query.semester];
-    db.query(getDatesQuery, getDatesParams)
-      .then((dates) => {
-        res.send(dates);
-      })
-      .catch((err) => {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.post(
-    "/editProject",
-    [
-      UserAuth.isAdmin,
-      UserAuth.canWrite,
-      // TODO: Should the max length be set to something smaller than 5000?
-      body("title")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 50 }),
-      body("organization")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("primary_contact")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("contact_email")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("contact_phone")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("background_info")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_description")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_scope")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("project_challenges")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("constraints_assumptions")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("sponsor_provided_resources")
-        .trim()
-        .escape()
-        .isLength({ max: 5000 }),
-      body("project_search_keywords").trim().escape().isLength({ max: 5000 }),
-      body("sponsor_deliverables")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-      body("proprietary_info").trim().escape().isLength({ max: 5000 }),
-      body("sponsor_alternate_time").trim().escape().isLength({ max: 5000 }),
-      body("sponsor_avail_checked")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("project_agreements_checked")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("assignment_of_rights")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty")
-        .isLength({ max: 5000 }),
-
-      body("team_name")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("poster")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("video")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("website")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("synopsis")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("sponsor")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      body("semester")
-        .not()
-        .isEmpty()
-        .trim()
-        .escape()
-        .withMessage("Cannot be empty"),
-      // body("date").not().isEmpty().trim().escape().withMessage("Cannot be empty"),
-      body("projectCoaches").trim().escape().isLength({ max: 5000 }),
-    ],
-    async (req, res, next) => {
-      let body = req.body;
-
-      const updateProjectSql = `UPDATE ${DB_CONFIG.tableNames.senior_projects}
-            SET status=?, title=?, display_name=?, organization=?, primary_contact=?, contact_email=?, contact_phone=?,
-            background_info=?, project_description=?, project_scope=?, project_challenges=?,
-            sponsor_provided_resources=?, project_search_keywords=?, constraints_assumptions=?, sponsor_deliverables=?,
-            proprietary_info=?, sponsor_alternate_time=?, sponsor_avail_checked=?, project_agreements_checked=?, assignment_of_rights=?,
-            team_name=?, poster=?, video=?, website=?, synopsis=?, sponsor=?, semester=?
-            WHERE project_id = ?`;
-
-      const updateProjectParams = [
-        body.status,
-        body.title,
-        body.display_name ? body.display_name : null, // Empty strings should be turned to null
-        body.organization,
-        body.primary_contact,
-        body.contact_email,
-        body.contact_phone,
-        body.background_info,
-        body.project_description,
-        body.project_scope,
-        body.project_challenges,
-        body.sponsor_provided_resources,
-        body.project_search_keywords,
-        body.constraints_assumptions,
-        body.sponsor_deliverables,
-        body.proprietary_info,
-        body.sponsor_alternate_time,
-        body.sponsor_avail_checked,
-        body.project_agreements_checked,
-        body.assignment_of_rights,
-        body.team_name,
-        body.poster,
-        body.video,
-        body.website,
-        body.synopsis,
-        body.sponsor,
-        body.semester || null,
-        body.project_id,
-      ];
-
-      const insertValues = body.projectCoaches
-        .split(",")
-        .map((coachId) => ` ('${body.project_id}', '${coachId}')`);
-      const deleteValues = body.projectCoaches
-        .split(",")
-        .map((coachId) => `'${coachId}'`);
-      const updateCoachesSql = `INSERT OR IGNORE INTO '${DB_CONFIG.tableNames.project_coaches}' ('project_id', 'coach_id') VALUES ${insertValues};`;
-      const deleteCoachesSQL = `DELETE FROM '${DB_CONFIG.tableNames.project_coaches}'
-                                        WHERE project_coaches.project_id = '${body.project_id}'
-                                        AND project_coaches.coach_id NOT IN (${deleteValues});`;
-
-      Promise.all([
-        db.query(updateProjectSql, updateProjectParams),
-        db.query(updateCoachesSql),
-        db.query(deleteCoachesSQL),
-      ])
-        .then((values) => {
-          return res.sendStatus(200);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
-          error.statusCode = 500;
-          return next(error);
-        });
-    },
-  );
-
-  /**
-   * Updates a proposal with the given information
-   */
-  db_router.patch(
-    "/updateProposalStatus",
-    [
-      UserAuth.isAdmin,
-      UserAuth.canWrite,
-      body("*").trim().escape().isJSON().isAlphanumeric(),
-    ],
-    (req, res, next) => {
-      const query = `UPDATE ${DB_CONFIG.tableNames.senior_projects} SET status = ? WHERE project_id = ?`;
-      db.query(query, [req.body.status, req.body.project_id])
-        .then(() => {
-          res.sendStatus(200);
-        })
-        .catch((err) => {
-          console.error(err);
-          const error = new Error(err);
           error.statusCode = 500;
           return next(error);
         });
@@ -2335,33 +1215,6 @@ module.exports = (db) => {
         });
     },
   );
-
-  db_router.get("/getArchivePoster", (req, res) => {
-    res.sendFile(
-      path.join(
-        __dirname,
-        "../../resource/archivePosters/" + req.query.fileName,
-      ),
-    );
-  });
-
-  db_router.get("/getArchiveVideo", (req, res) => {
-    res.sendFile(
-      path.join(
-        __dirname,
-        "../../resource/archiveVideos/" + req.query.fileName,
-      ),
-    );
-  });
-
-  db_router.get("/getArchiveImage", (req, res) => {
-    res.sendFile(
-      path.join(
-        __dirname,
-        "../../resource/archiveImages/" + req.query.fileName,
-      ),
-    );
-  });
 
   db_router.get(
     "/getActiveTimelines",
@@ -3025,20 +1878,6 @@ module.exports = (db) => {
     },
   );
 
-  /**
-   * This is for getting the archive data to view/edit based on a specific ID.
-   * */
-  db_router.get("/getArchiveProject", [UserAuth.isAdmin], (req, res) => {
-    let query = `
-            SELECT *
-            FROM archive
-            WHERE archive_id = ?
-        `;
-
-    const params = [req.query.archive_id];
-    db.query(query, params).then((project) => res.send(project));
-  });
-
   db_router.get("/getSubmission", [UserAuth.isSignedIn], (req, res, next) => {
     let getSubmissionQuery = "";
     let params = [];
@@ -3203,196 +2042,6 @@ module.exports = (db) => {
     },
   );
 
-  db_router.get("/searchForArchive", (req, res, next) => {
-    const { resultLimit, offset, searchQuery, inactive } = req.query;
-    let skipNum = offset * resultLimit;
-    let getProjectsQuery = "";
-    let queryParams = [];
-    let getProjectsCount = "";
-    let projectCountParams = [];
-
-    // allow inactive projects in search
-    if (inactive === "true") {
-      getProjectsQuery = `SELECT * FROM  archive WHERE
-                            archive.OID NOT IN (
-                SELECT OID
-                FROM archive
-                WHERE title like ?
-                   OR sponsor like ?
-                   OR members like ?
-                   OR coach like ?
-                   OR keywords like ?
-                   OR synopsis like ?
-                   OR url_slug like ?
-                ORDER BY title,
-                         sponsor,
-                         members,
-                         coach,
-                         keywords,
-                         synopsis,
-                         url_slug
-            LIMIT ?
-            ) AND (
-                archive.title like ?
-                OR archive.sponsor like ?
-                OR archive.members like ?
-                OR archive.coach like ?
-                OR archive.keywords like ?
-                OR archive.synopsis like ?
-                OR archive.url_slug like ?
-                )
-            ORDER BY
-                archive.title,
-                archive.sponsor,
-                archive.members,
-                archive.coach,
-                archive.keywords,
-                archive.synopsis,
-                archive.url_slug
-            LIMIT ?`;
-
-      getProjectsCount = `SELECT COUNT(*)
-                            FROM archive
-                            WHERE
-                                title like ?
-                                OR sponsor like ?
-                                OR members like ?
-                                OR coach like ?
-                                OR keywords like ?
-                                OR synopsis like ?
-                                OR url_slug like ?
-                                `;
-    } else {
-      getProjectsQuery = `SELECT * FROM  archive WHERE
-                            archive.OID NOT IN (
-                SELECT OID
-                FROM archive
-                WHERE title like ?
-                   OR sponsor like ?
-                   OR members like ?
-                   OR coach like ?
-                   OR keywords like ?
-                   OR synopsis like ?
-                   OR url_slug like ?
-                   AND inactive = ''
-                ORDER BY title,
-                         sponsor,
-                         members,
-                         coach,
-                         keywords,
-                         synopsis,
-                         url_slug
-            LIMIT ?
-            ) AND (
-                archive.title like ?
-                OR archive.sponsor like ?
-                OR archive.members like ?
-                OR archive.coach like ?
-                OR archive.keywords like ?
-                OR archive.synopsis like ?
-                OR archive.url_slug like ?
-                AND inactive = ''
-                )
-            ORDER BY
-                archive.title,
-                archive.sponsor,
-                archive.members,
-                archive.coach,
-                archive.keywords,
-                archive.synopsis,
-                archive.url_slug
-            LIMIT ?`;
-
-      getProjectsCount = `SELECT COUNT(*)
-                            FROM archive
-                            WHERE
-                                title like ?
-                                OR sponsor like ?
-                                OR members like ?
-                                OR coach like ?
-                                OR keywords like ?
-                                OR synopsis like ?
-                                OR url_slug like ?
-                                AND (inactive = '' OR inactive IS NULL)
-                                `;
-    }
-
-    const searchQueryParam = searchQuery || "";
-
-    queryParams = [
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      skipNum || 0,
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      resultLimit || 0,
-    ];
-    projectCountParams = [
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-      "%" + searchQueryParam + "%",
-    ];
-
-    const projectPromise = db.query(getProjectsQuery, queryParams);
-    const projectCountPromise = db.query(getProjectsCount, projectCountParams);
-    Promise.all([projectCountPromise, projectPromise])
-      .then(([[projectCount], projectRows]) => {
-        res.send({
-          projectCount: projectCount[Object.keys(projectCount)[0]],
-          projects: projectRows,
-        });
-      })
-      .catch((err) => {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.get("/getArchiveFromSlug", (req, res, next) => {
-    let query = `SELECT * FROM archive WHERE url_slug=?`;
-    let params = [req.query.url_slug];
-    db.query(query, params)
-      .then((values) => {
-        res.status(200).send(values);
-      })
-      .catch((err) => {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.get("/getArchiveFromProject", (req, res) => {
-    let query = `SELECT * FROM archive WHERE archive.project_id=?`;
-    let params = [req.query.project_id];
-    db.query(query, params)
-      .then((values) => {
-        res.status(200).send(values);
-      })
-      .catch((err) => {
-        console.error(err);
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
   db_router.post(
     "/createAction",
     [UserAuth.isAdmin, UserAuth.canWrite, body("page_html").unescape()],
@@ -3444,21 +2093,6 @@ module.exports = (db) => {
             ORDER BY end_date, start_date, name
         `;
     db.query(getSemestersQuery)
-      .then((values) => {
-        res.send(values);
-      })
-      .catch((err) => {
-        const error = new Error(err);
-        error.statusCode = 500;
-        return next(error);
-      });
-  });
-
-  db_router.get("/getArchive", [UserAuth.isAdmin], (req, res, next) => {
-    let getArchiveQuery = `
-            SELECT *
-            FROM archive`;
-    db.query(getArchiveQuery)
       .then((values) => {
         res.send(values);
       })
